refector search and filters
