+++ src/ui/layouts/connectors/ConnectorDetail/index.tsx
@@ -15,15 +15,19 @@ import { DEFAULT_WORKSPACE_NAME } from '../../../../constants';
 import { CollapseTable } from '../../common/CollapseTable';
 import { GetHeaderCols } from './getHeaderCols';
 import { ConnectorComponents } from './ConnectorComponents';
-import FilterComponent, {
-  getInitialFilterState,
-} from '../../../components/Filters';
+import FilterComponent from '../../../components/Filters';
+import {
+  getInitialFilterStateStackComponents,
+  searchParamConstants,
+} from '../../stackComponents/Stacks/filterParamConstants';
 
 const FilterWrapper = () => {
   const locationPath = useLocationPath();
 
   // TODO: Dev please note: getInitialFilterState is for stack inital filter value for any other component you need to modify it
-  const [filters, setFilter] = useState([getInitialFilterState()]);
+  const [filters, setFilter] = useState([
+    getInitialFilterStateStackComponents(),
+  ]);
   function getFilter(values: any) {
     const filterValuesMap = values.map((v: any) => {
       return {
@@ -37,7 +41,8 @@ const FilterWrapper = () => {
   return (
     <Box style={{ marginTop: '10px', width: '100%' }}>
       <FilterComponent
-        getInitials={getInitialFilterState}
+        searchColumns={searchParamConstants}
+        getInitials={getInitialFilterStateStackComponents}
         filters={filters}
         setFilter={setFilter}
       >
